refactor: config  SSR with redux
