+++ components/CryptoCurrenciesTable/CryptoCurrenciesTable.tsx
