+++ components/CryptoCurrenciesTable/CryptoCurrenciesTableSkeleton.tsx
