+++ lib/currentPage.ts
