+++ lib/types/CryptoCurrency.ts
