+++ package.json
@@ -19,6 +19,7 @@
     "@reduxjs/toolkit": "^1.9.7",
     "clsx": "^2.0.0",
     "next": "14.0.3",
+    "next-redux-wrapper": "^8.1.0",
     "react": "^18",
     "react-dom": "^18",
     "react-redux": "^8.1.3"
