+++ pages/_app.tsx
@@ -1,0 +1,31 @@
+import type { ReactElement, ReactNode } from 'react'
+import type { NextPage } from 'next'
+import type { AppProps } from 'next/app'
+import { wrapper } from '../lib/store'
+import { Provider } from 'react-redux'
+import { config } from '@fortawesome/fontawesome-svg-core'
+import Layout from '@/components/Layout/Layout'
+import '@/style/globals.css'
+config.autoAddCss = false
+
+
+
+export type NextPageWithLayout<P = {}, IP = P> = NextPage<P, IP> & {
+  getLayout?: (page: ReactElement) => ReactNode
+}
+
+type AppPropsWithLayout = AppProps & {
+  Component: NextPageWithLayout
+}
+
+export default function MyApp({ Component, pageProps }: AppPropsWithLayout) {
+  const { store, props } = wrapper.useWrappedStore(pageProps)
+
+  return (
+    <Provider store={store}>
+      <Layout>
+        <Component {...pageProps} />
+      </Layout>
+    </Provider>
+  )
+}
