--- src/app/__mocks__/cryptoCurrencies.ts
@@ -1,76 +1,0 @@
-import { CryptoCurrency } from "../lib/types/CryptoCurrency"
-
-export const cryptoCurrenciesMock: CryptoCurrency[] = [
-  {
-    id: "90",
-    symbol: "BTC",
-    name: "Bitcoin",
-    nameid: "bitcoin",
-    rank: 1,
-    price_usd: "37689.47",
-    percent_change_24h: "1.29",
-    percent_change_1h: "0.02",
-    percent_change_7d: "3.18",
-    price_btc: "1.00",
-    market_cap_usd: "733849703359.39",
-    volume24: 15628323625.924067,
-    volume24a: 19029531229.4653,
-    csupply: "19470946.00",
-    tsupply: "19470946",
-    msupply: "21000000"
-  },
-  {
-    id: "80",
-    symbol: "ETH",
-    name: "Ethereum",
-    nameid: "ethereum",
-    rank: 2,
-    price_usd: "2102.75",
-    percent_change_24h: "2.20",
-    percent_change_1h: "0.08",
-    percent_change_7d: "5.93",
-    price_btc: "0.055839",
-    market_cap_usd: "257325154135.38",
-    volume24: 9073224807.988739,
-    volume24a: 10802347965.53811,
-    csupply: "122375302.00",
-    tsupply: "122375302",
-    msupply: ""
-  },
-  {
-    id: "518",
-    symbol: "USDT",
-    name: "Tether",
-    nameid: "tether",
-    rank: 3,
-    price_usd: "1.00",
-    percent_change_24h: "-0.09",
-    percent_change_1h: "-0.02",
-    percent_change_7d: "0.04",
-    price_btc: "0.000027",
-    market_cap_usd: "83747230252.66",
-    volume24: 32591087527.725147,
-    volume24a: 39544899919.43293,
-    csupply: "83638374850.00",
-    tsupply: "86715788906.061",
-    msupply: ""
-  },
-  {
-    id: "2710",
-    symbol: "BNB",
-    name: "Binance Coin",
-    nameid: "binance-coin",
-    rank: 4,
-    price_usd: "234.13",
-    percent_change_24h: "0.62",
-    percent_change_1h: "-0.13",
-    percent_change_7d: "-4.50",
-    price_btc: "0.006217",
-    market_cap_usd: "39052970654.27",
-    volume24: 312346707.486959,
-    volume24a: 508035492.5248391,
-    csupply: "166801148.00",
-    tsupply: "192443301",
-    msupply: "200000000"
-  }
-];
